+++ src/index.html
@@ -234,6 +234,9 @@
                 <p class="text-person">"I'm so happy because i found this recipe, and it just made my life easier.<br> Thanks
                     so much for sharing!"</p>
             </div>
+            <div class="p-sp-person">
+                    <span class="sp-person">-Michael Dawson, San Francisco, CA-</span>
+            </div>
         </div>
     </div>
     <div class="p-slider">
@@ -245,6 +248,9 @@
             <div class="p-text-person">
                 <p class="text-person">"I'm so happy because i found this recipe, and it just made my life easier.<br> Thanks
                     so much for sharing!"</p>
+            </div>
+            <div class="p-sp-person">
+                    <span class="sp-person">-Michael Dawson, San Francisco, CA-</span>
             </div>
 
         </div>
@@ -259,10 +265,96 @@
                 <p class="text-person">"I'm so happy because i found this recipe, and it just made my life easier.<br> Thanks
                     so much for sharing!"</p>
             </div>
+            <div class="p-sp-person">
+                    <span class="sp-person">-Michael Dawson, San Francisco, CA-</span>
+            </div>
 
         </div>
     </div>
 </section> 
+
+    <section class="cuisines-container">
+    <div class="p-cuisines">
+        <div class="container">
+            <div class="row">
+                <div class="col-lg-8 offset-lg-5">
+                    <div class="p-title-main">
+                        <h3 class="heading-3">
+                            browse by cuisines
+                        </h3>
+                    </div>
+
+                </div>
+            </div>
+        </div>
+
+
+        <div class="p-img">
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-1.png');">
+                <div class="p-title">
+                    <span class="sp-eat">327 Recipes</span>
+                    <h4 class="heading-4">ITALIA</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-2.png');">
+                <div class="p-title">
+                    <span class="sp-eat">856 Recipes</span>
+                    <h4 class="heading-4">INDIAN</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-3.png');">
+                <div class="p-title">
+                    <span class="sp-eat">27 Recipes</span>
+                    <h4 class="heading-4">FRENCH</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-4.png');">
+                <div class="p-title">
+                    <span class="sp-eat">174 Recipes</span>
+                    <h4 class="heading-4">STEAKHOUSE</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-5.png');">
+                <div class="p-title">
+                    <span class="sp-eat">737 Recipes</span>
+                    <h4 class="heading-4">SEAFOOD</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-6.png');">
+                <div class="p-title">
+                    <span class="sp-eat">327 Recipes</span>
+                    <h4 class="heading-4">SUSHI</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-7.png');">
+                <div class="p-title">
+                    <span class="sp-eat">527 Recipes</span>
+                    <h4 class="heading-4">MEXICAN</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-8.png');">
+                <div class="p-title">
+                    <span class="sp-eat">145 Recipes</span>
+                    <h4 class="heading-4">CHINESE</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-9.png');">
+                <div class="p-title">
+                    <span class="sp-eat">327 Recipes</span>
+                    <h4 class="heading-4">PIZZA</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-10.png');">
+                <div class="p-title">
+                    <span class="sp-eat">1347 Recipes</span>
+                    <h4 class="heading-4">AMERICAN</h4>
+                </div>
+            </div>
+
+
+        </div>
+    </div>
+</section>
     <!-- include footer.html -->
     <footer class="footer-container">
   <div class="container">
@@ -290,6 +382,7 @@
 <script type="text/javascript" src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
 <!-- endbuild -->
 <!-- build:js script/script.min.js -->
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 <script type="text/javascript" src="scripts/script.js"></script>
 <script type="text/javascript" src="../../scripts/slider-person.js"></script>
 <script type="text/javascript" src="../../scripts/slick.js"></script>
