+++ src/styles/layout/_slider-person-them.scss
@@ -67,7 +67,7 @@ $slick-opacity-not-active: 0.25 !default;
     height: 20px;
     width: 20px;
     line-height: 0px;
-    font-size: 0px;
+    @include rem("font-size", 0);
     cursor: pointer;
     background: transparent;
     color: transparent;
@@ -130,13 +130,13 @@ $slick-opacity-not-active: 0.25 !default;
 
 /* Dots */
 
-.slick-dotted.slick-slider {
-    margin-bottom: 30px;
-}
+// .slick-dotted.slick-slider {
+    
+// }
 
 .slick-dots {
     position: absolute;
-    bottom: -25px;
+    bottom: 25px;
     list-style: none;
     display: block;
     text-align: center;
@@ -159,7 +159,7 @@ $slick-opacity-not-active: 0.25 !default;
             width: 20px;
             outline: none;
             line-height: 0px;
-            font-size: 0px;
+            @include rem("font-size", 0);
             color: transparent;
             padding: 5px;
             cursor: pointer;
@@ -176,8 +176,8 @@ $slick-opacity-not-active: 0.25 !default;
                 content: $slick-dot-character;
                 width: 20px;
                 height: 20px;
-                font-family: $slick-font-family;
-                font-size: $slick-dot-size;
+                font-family:'Lato';
+                @include rem("font-size", 25);
                 line-height: 20px;
                 text-align: center;
                 color: $slick-dot-color;
@@ -187,7 +187,7 @@ $slick-opacity-not-active: 0.25 !default;
             }
         }
         &.slick-active button:before {
-            color: $slick-dot-color-active;
+            color: $cl-style;
             opacity: $slick-opacity-default;
         }
     }
