+++ src/styles/layout/_slider.scss
@@ -5,7 +5,7 @@
     height: 80vh;
     .banner {
       width: 100%;
-      height: 80vh;
+      height: 100%;
     }
     .carousel-caption {
       position: absolute;
