+++ src/styles/main.scss
@@ -37,6 +37,7 @@
 @import "layout/slider-person";
 @import "layout/slider-person-them";
 @import "layout/slick";
+@import "layout/cuisines";
 @import "layout/footer";
 
 // ----------------------------------------------------------------------
