+++ src/views/layouts/scripts.html
@@ -3,6 +3,7 @@
 <script type="text/javascript" src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
 <!-- endbuild -->
 <!-- build:js script/script.min.js -->
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 <script type="text/javascript" src="scripts/script.js"></script>
 <script type="text/javascript" src="../../scripts/slider-person.js"></script>
 <script type="text/javascript" src="../../scripts/slick.js"></script>
