+++ src/views/pages/index.html
@@ -25,6 +25,8 @@
     @@include('../partials/slider.html')
 
     @@include('../partials/slider-person.html') 
+
+    @@include('../partials/cuisines.html')
     <!-- include footer.html -->
     @@include('../layouts/footer.html')
   </div>
