+++ src/views/partials/cuisines.html
@@ -1,0 +1,82 @@
+<section class="cuisines-container">
+    <div class="p-cuisines">
+        <div class="container">
+            <div class="row">
+                <div class="col-lg-8 offset-lg-5">
+                    <div class="p-title-main">
+                        <h3 class="heading-3">
+                            browse by cuisines
+                        </h3>
+                    </div>
+
+                </div>
+            </div>
+        </div>
+
+
+        <div class="p-img">
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-1.png');">
+                <div class="p-title">
+                    <span class="sp-eat">327 Recipes</span>
+                    <h4 class="heading-4">ITALIA</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-2.png');">
+                <div class="p-title">
+                    <span class="sp-eat">856 Recipes</span>
+                    <h4 class="heading-4">INDIAN</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-3.png');">
+                <div class="p-title">
+                    <span class="sp-eat">27 Recipes</span>
+                    <h4 class="heading-4">FRENCH</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-4.png');">
+                <div class="p-title">
+                    <span class="sp-eat">174 Recipes</span>
+                    <h4 class="heading-4">STEAKHOUSE</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-5.png');">
+                <div class="p-title">
+                    <span class="sp-eat">737 Recipes</span>
+                    <h4 class="heading-4">SEAFOOD</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-6.png');">
+                <div class="p-title">
+                    <span class="sp-eat">327 Recipes</span>
+                    <h4 class="heading-4">SUSHI</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-7.png');">
+                <div class="p-title">
+                    <span class="sp-eat">527 Recipes</span>
+                    <h4 class="heading-4">MEXICAN</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-8.png');">
+                <div class="p-title">
+                    <span class="sp-eat">145 Recipes</span>
+                    <h4 class="heading-4">CHINESE</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-9.png');">
+                <div class="p-title">
+                    <span class="sp-eat">327 Recipes</span>
+                    <h4 class="heading-4">PIZZA</h4>
+                </div>
+            </div>
+            <div class="image-eat image-1" style="background-image: url('../images/eat-item-10.png');">
+                <div class="p-title">
+                    <span class="sp-eat">1347 Recipes</span>
+                    <h4 class="heading-4">AMERICAN</h4>
+                </div>
+            </div>
+
+
+        </div>
+    </div>
+</section>
