+++ src/views/partials/slider-person.html
@@ -9,6 +9,9 @@
                 <p class="text-person">"I'm so happy because i found this recipe, and it just made my life easier.<br> Thanks
                     so much for sharing!"</p>
             </div>
+            <div class="p-sp-person">
+                    <span class="sp-person">-Michael Dawson, San Francisco, CA-</span>
+            </div>
         </div>
     </div>
     <div class="p-slider">
@@ -20,6 +23,9 @@
             <div class="p-text-person">
                 <p class="text-person">"I'm so happy because i found this recipe, and it just made my life easier.<br> Thanks
                     so much for sharing!"</p>
+            </div>
+            <div class="p-sp-person">
+                    <span class="sp-person">-Michael Dawson, San Francisco, CA-</span>
             </div>
 
         </div>
@@ -33,6 +39,9 @@
             <div class="p-text-person">
                 <p class="text-person">"I'm so happy because i found this recipe, and it just made my life easier.<br> Thanks
                     so much for sharing!"</p>
+            </div>
+            <div class="p-sp-person">
+                    <span class="sp-person">-Michael Dawson, San Francisco, CA-</span>
             </div>
 
         </div>
